+++ apps/docs/src/app/core/component-docs/object-marker/object-marker-header/object-marker-header.component.ts
@@ -1,0 +1,7 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+    selector: 'fd-object-marker-header',
+    templateUrl: './object-marker-header.component.html'
+})
+export class ObjectMarkerHeaderComponent {}