+++ apps/docs/src/app/core/component-docs/shellbar/examples/shellbar-side-nav/shellbar-side-nav-example.component.ts
@@ -1,9 +1,10 @@
-import { Component } from '@angular/core';
+import { Component, ViewEncapsulation } from '@angular/core';
 
 @Component({
     selector: 'fd-shellbar-side-nav-example',
     templateUrl: './shellbar-side-nav-example.component.html',
-    styleUrls: ['./shellbar-side-nav-example.component.scss']
+    styleUrls: ['./shellbar-side-nav-example.component.scss'],
+    encapsulation: ViewEncapsulation.None
 })
 export class ShellbarSideNavExampleComponent {
     condensed = false;