+++ libs/core/src/lib/nested-list/nested-list-directives.ts
@@ -1,7 +1,22 @@
-import { Directive, ElementRef, HostBinding, HostListener, Input, OnChanges, OnInit } from '@angular/core';
+import {
+    ChangeDetectionStrategy,
+    ChangeDetectorRef,
+    Component,
+    Directive,
+    ElementRef,
+    HostBinding,
+    HostListener,
+    Input,
+    OnChanges,
+    OnInit, Optional,
+    ViewEncapsulation
+} from '@angular/core';
 import { NestedItemService } from './nested-item/nested-item.service';
 import { applyCssClass } from '../utils/decorators/apply-css-class.decorator';
 import { CssClassBuilder } from '../utils/interfaces/css-class-builder.interface';
+import { RtlService } from '../utils/services/rtl.service';
+import { map } from 'rxjs/operators';
+import { Observable, of } from 'rxjs';
 
 @Directive({
     selector: '[fdNestedDirectivesHeader], [fd-nested-list-header]'
@@ -88,14 +103,21 @@ export class NestedListTitleDirective {
     }
 }
 
-@Directive({
+@Component({
+    // tslint:disable-next-line:component-selector
     selector: '[fdNestedListExpandIcon], [fd-nested-list-expand-icon]',
+    template: `
+        <ng-content></ng-content>
+        <fd-icon [glyph]="expanded ? 'navigation-down-arrow' : ( sideArrowIcon$ | async )"></fd-icon>
+    `,
     host: {
         'aria-haspopup': 'true',
         'tabindex': '-1'
-    }
+    },
+    changeDetection: ChangeDetectionStrategy.OnPush,
+    encapsulation: ViewEncapsulation.None
 })
-export class NestedListExpandIconDirective {
+export class NestedListExpandIconComponent {
 
     /** @hidden */
     @HostBinding('class.fd-nested-list__button')
@@ -113,9 +135,16 @@ export class NestedListExpandIconDirective {
     @HostBinding('attr.aria-expanded')
     expanded = false;
 
+    /** @hidden */
+    sideArrowIcon$: Observable<string>;
+
     constructor (
-        private _itemService: NestedItemService
-    ) {}
+        private _itemService: NestedItemService,
+        private _changeDetRef: ChangeDetectorRef,
+        @Optional() private _rtlService: RtlService
+    ) {
+        this._listenOnTextDirection();
+    }
 
     /** Mouse event handler */
     @HostListener('click', ['$event'])
@@ -129,5 +158,18 @@ export class NestedListExpandIconDirective {
     @HostListener('focus')
     onFocus(): void {
         this._itemService.focus.next();
+    }
+
+    /** @hidden */
+    changeExpandedState(expanded: boolean): void {
+        this.expanded = expanded;
+        this._changeDetRef.detectChanges();
+    }
+
+    /** @hidden Sets expand arrow depending on text direction */
+    private _listenOnTextDirection(): void {
+        this.sideArrowIcon$ = this._rtlService
+            ? this._rtlService.rtl.pipe(map(isRtl => isRtl ? 'navigation-left-arrow' : 'navigation-right-arrow'))
+            : of('navigation-right-arrow');
     }
 }