+++ libs/core/src/lib/object-marker/public_api.ts
@@ -1,0 +1,2 @@
+export * from './object-marker.module';
+export * from './object-marker.component';