+++ libs/platform/src/lib/components/button/button.component.ts
@@ -1,22 +1,26 @@
-import { Component, Input, Output, EventEmitter, ElementRef, ChangeDetectorRef, AfterViewInit, HostBinding } from '@angular/core';
+import {
+    Component, Input, Output, EventEmitter, ElementRef,
+    ChangeDetectorRef, AfterViewInit, HostBinding
+} from '@angular/core';
+import { GlyphPosition, ButtonType } from '@fundamental-ngx/core';
 import { BaseComponent } from '../base';
 
-export type ButtonType =
-    | ''
-    | 'standard'
-    | 'positive'
-    | 'negative'
-    | 'attention'
-    | 'ghost'
-    | 'transparent'
-    | 'emphasized'
-    | 'menu';
 @Component({
     selector: 'fdp-button',
     templateUrl: './button.component.html',
     styleUrls: ['./button.component.scss']
 })
 export class ButtonComponent extends BaseComponent implements AfterViewInit {
+
+    /** Position of glyph related to text */
+    @Input()
+    glyphPosition: GlyphPosition = 'before';
+
+    /**
+    * Text rendered inside button component
+    */
+    @Input()
+    label: string;
 
     /** The icon to include in the button. See the icon page for the list of icons.
      * Setter is used to control when css class have to be rebuilded.
@@ -30,7 +34,7 @@ export class ButtonComponent extends BaseComponent implements AfterViewInit {
      'transparent', 'emphasized','menu'.
      *Leave empty for default (standard button).'*/
     @Input()
-    buttonType: ButtonType;
+    buttonType: ButtonType = 'standard';
 
     /** arialabel, tooltip for truncated text
      * for acccesiblity of the element */